+++ HISTORY.md
@@ -49,6 +49,7 @@
 
 ### 2025-12-11
 * objects, extensions: added support for pen modes ("paint", "erase" and "create")
+* objects: support stamping on sprites using pen modes
 
 ### 2025-12-10
 * gui: tweaked IDE resizing layout for different translations